requirements.txt
streamlit
pandas
seaborn
matplotlib.pyplot
numpy
scikit-learn